[TASK] Avoid mocking of nodes in tests (#1143)
Backport of test adjustments in Fluid v5.
